+++ kes-agent/src/Cardano/KESAgent/Serialization/CBOR.hs
@@ -17,6 +17,7 @@ import Cardano.KESAgent.KES.OCert
 import Cardano.KESAgent.Serialization.TextEnvelope
 
 import Cardano.Binary (FromCBOR (..), ToCBOR (..))
+import Cardano.Binary.FixedSizeCodec (decodeFixedSized, encodeFixedSized)
 import Cardano.Crypto.DSIGN.Class
 import Cardano.Crypto.KES.Class
 
@@ -96,10 +97,10 @@ instance
 newtype Sigma c = Sigma (SignedDSIGN (DSIGN c) (OCertSignable c))
 
 instance (Typeable c, Crypto c) => ToCBOR (Sigma c) where
-  toCBOR (Sigma x) = encodeSignedDSIGN x
+  toCBOR (Sigma x) = encodeFixedSized x
 
 instance (Typeable c, Crypto c) => FromCBOR (Sigma c) where
-  fromCBOR = Sigma <$> decodeSignedDSIGN
+  fromCBOR = Sigma <$> decodeFixedSized
 
 -- | The operational certificate, as serialized. The serialization format should
 -- be compatible with what @cardano-node@ uses.